Copied the source code from the SourceForge repo at https://sourceforge.net/p/opendx/code/HEAD/tree/ (taken on Mar 30, 2021), which appears to be a resqued source tree from berlios.de OpenDX2 repo.
This source tree carries version tag 4.4.5, and appears to be significantly
more recent than the 4.4.4 distribution we are based on. It incorporates a
number of fixes to the SMP support under Linux, and appears to have partial
support for large arenas. At the same time, it still contains many of the
problems fixed in our trunk.

The plan is to eventually fold these patches into our trunk.

+++ src/uipp/dxui/SequencerNode.h
@@ -9,10 +9,8 @@
 #include <dxconfig.h>
 #include "../base/defines.h"
 
-
 #ifndef _SequencerNode_h
 #define _SequencerNode_h
-
 
 #include "ShadowedOutputNode.h"
 
@@ -56,15 +54,9 @@ class SequencerNode : public ShadowedOutputNode
     int		    startValue;	    	// Start frame
     int		    stopValue;		// Stop frame
 
-    //short           minimum;        /* minimum frame                */
-    //short           maximum;        /* maximum frame                */
-    //short           start;          /* start frame                  */
-    //short           stop;           /* stop frame                   */
-    //short           increment;      /* frame increment              */
-
-    short           current;        /* current frame                */
-    short           next;           /* current frame                */
-    short           previous;       /* previous frame               */
+    int           current;        /* current frame                */
+    int           next;           /* current frame                */
+    int           previous;       /* previous frame               */
 
     boolean	    ignoreFirstFrameMsg;
     boolean         transmitted;    /* frame values transmitted?    */
